% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_example.R
\name{baraks_magnificent_function}
\alias{baraks_magnificent_function}
\title{Barak's magnificent function}
\usage{
baraks_magnificent_function(x, y)
}
\arguments{
\item{x}{description param1}

\item{y}{some other praram param2 must be numeric}
}
\value{
the product of \code{x} and \code{y}
}
\description{
This function multiplies two numbers
}
\details{
here we add a in depth description of the method
\eqn{H_0^{(u)}}
}
\examples{
baraks_magnificent_function(2,5)
}
